8328279: Convert java/awt/Cursor/CursorOverlappedPanelsTest test to main
Backport-of: 2094ff3c7ad829355bf13760d9a19b37649b3517
